+++ plugins/example-todo-list-backend/CHANGELOG.md
@@ -1,5 +1,20 @@
 # @internal/plugin-todo-list-backend
 
+## 1.0.46
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.6.0
+
+## 1.0.46-next.1
+
+### Patch Changes
+
+- Updated dependencies
+  - @backstage/backend-plugin-api@1.6.0-next.1
+  - @backstage/errors@1.2.7
+
 ## 1.0.46-next.0
 
 ### Patch Changes